Support several backports in a single command.
When working in large projects with more than one release branch the old
code made us write a comment PER branch, resulting in something like:

/sudo backport d16-9
/sudo backport xcode12.4
/sudo backport xcode12.5

That is a lot of work and developers are lazy animals. We already
supported using more than one branch in the regexp but it resulted in an
error later in the script. This commit makes the required changes to
ensure that we support several branches.

1. Pares all branches, creates a collection in the params.
2. If only one branhc is used, we create a collection with a single
   object, else several.
3. We trigger a backport PER branch in a loop using a scriptblock.
4. The exit code is the addition of all the execitions, resulting in 0
   if all worked or the number of errors.